requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
blinker==1.4
colorama==0.4.4
commonmark==0.9.1
cssselect==1.1.0
Django==4.0.2
django-autoslug==1.9.8
django-tinymce==3.4.0
docutils==0.18.1
feedgenerator==2.0.0
Jinja2==3.0.3
lxml==4.7.1
markdown2==2.4.2
MarkupSafe==2.0.1
mysqlclient==2.1.0
pelican==4.7.2
Pillow==9.0.1
Pygments==2.11.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
readtime==1.1.1
rich==11.2.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
Unidecode==1.3.2
